+++ src/plugin/tasks/file-exists.ts
@@ -1,0 +1,5 @@
+import { existsSync } from 'fs';
+
+import { defineTask } from '../lib/utils';
+
+export default defineTask((path: string) => existsSync(path));