+++ src/support/commands/crud.ts
@@ -1,8 +1,9 @@
 import { normalizeSparql } from '@noeldemartin/solid-utils';
 
+import { applyReplacements } from '../lib/replacements';
+import { fixtureOrText } from '../lib/fixtures';
 import { podUrl } from '../../shared';
-import { applyReplacements } from '../lib/utils';
-import type { Replacements } from '../lib/utils';
+import type { Replacements } from '../lib/replacements';
 
 export function solidCreateContainer(path: string, name: string = 'Container'): void {
     const containerUrl = podUrl(path);
@@ -23,8 +24,8 @@ export function solidCreateContainer(path: string, name: string = 'Container'): 
     });
 }
 
-export function solidCreateDocument(path: string, fixture: string, replacements: Replacements = {}): void {
-    cy.fixture(fixture).then((body) =>
+export function solidCreateDocument(path: string, turtleOrFixture: string, replacements: Replacements = {}): void {
+    fixtureOrText(turtleOrFixture).then((body) =>
         cy.solidRequest(podUrl(path), {
             method: 'PUT',
             headers: { 'Content-Type': 'text/turtle' },
@@ -40,8 +41,8 @@ export function solidReadDocument(path: string): Cypress.Chainable<string> {
     return cy.solidRequest(podUrl(path)).then((response) => response.body);
 }
 
-export function solidUpdateDocument(path: string, fixture: string, replacements: Replacements = {}): void {
-    cy.fixture(fixture).then((body) =>
+export function solidUpdateDocument(path: string, sparqlOrFixture: string, replacements: Replacements = {}): void {
+    fixtureOrText(sparqlOrFixture).then((body) =>
         cy.solidRequest(podUrl(path), {
             method: 'PATCH',
             headers: { 'Content-Type': 'application/sparql-update' },