+++ src/support/commands/fixtures.ts
@@ -1,6 +1,6 @@
-import { applyReplacements } from '../lib/utils';
-import type { Replacements } from '../lib/utils';
+import { applyReplacements } from '../lib/replacements';
+import type { Replacements } from '../lib/replacements';
 
-export function fixtureWithReplacements(filePath: string, replacements: Replacements): Cypress.Chainable<string> {
-    return cy.fixture(filePath).then((contents) => applyReplacements(contents, replacements));
+export function fixtureWithReplacements(path: string, replacements: Replacements): Cypress.Chainable<string> {
+    return cy.fixture(path).then((contents) => applyReplacements(contents, replacements));
 }